chore(app): consolidate admin commands + configure pages
